If anyone needs me, I'll be under the shame-rock
To my reviewer:

Last night I made what felt like my first breakthrough in the server-side testing. But it was more piecemeal than birds eye view. I have several blind spots here and I think with a better focused environment (I'm flying in several hours and had to pack and finish prepping myself, my house, and my car (dropping at the dealer) for flight). This week was in a surprising way some of the most fun I've had in awhile, so I'll certainly be coming back to this in a better head space.