Added Error Tests to Methods
server.test.js
- Fixed POST error test so that data posted would cause an error because
by excluding genre which is a required field, it could force an error
to occur which the test is able to check for

- Removed error test from GET. Does not seem necessary to have error test
on this method

- Added error test on PUT method by excluding the title which is needed
in order for proper functionality of this method. Therefore, an error is
sent back and the test can check that there is indeed a status of 422

- Added error test on DELETE method by sending the request to an ID that does
not exist. This forces the API to send back an error response which can be
checked by the test to make sure the status is 422